Correct case of Extension type
Changed the type of the ‘extension’ nodes I was creating to ‘Extension’, since type names should be Pascal case. This means the type is actually applied, instead of being defined but never applied to the nodes I was creating.

This rippled through quite a few places in the codebase. The main effect of this is that all the schema I was defining actually are applied to the extension nodes.

A code path that wasn’t being traversed, for reasons I couldn’t figure out, now is being traversed. It turned out to not be a great way of doing the thing, so I removed it. I also removed a bit of clutter from the starter.