Use new-style maven central URLs
It seems central.sonatype is less stable, so we get more 503s and spurious reports of URLs not existing, so I have removed the static-generation link checking and left it to the integration test.

In the integration test, we need to be less concurrent and retry more often in links test, to work around tendency of central.sonatype to return 503s.